Extract logic to helper class
